:white_check_mark: [#4246] Add tests for hashing of additional identifying attributes
